+++ src/routes/v2/admin/users/+page.svelte
@@ -50,7 +50,6 @@
 			<th>Active</th>
 			<th>Superuser</th>
 			<th>Verified</th>
-			<th>Slurm user</th>
 			<th>Actions</th>
 		</tr>
 	</thead>
@@ -64,7 +63,6 @@
 					<td><BooleanIcon value={user.is_active} /></td>
 					<td><BooleanIcon value={user.is_superuser} /></td>
 					<td><BooleanIcon value={user.is_verified} /></td>
-					<td>{user.slurm_user || '-'}</td>
 					<td>
 						<a href="/v2/admin/users/{user.id}" class="btn btn-light">
 							<i class="bi-info-circle" /> Info