+++ src/routes/v2/admin/users/[userId]/+page.svelte
@@ -5,8 +5,12 @@
 
 	/** @type {import('$lib/types').User & {group_ids: number[]}} */
 	const user = $page.data.user;
+	/** @type {import('$lib/types').UserSettings} */
+	const settings = $page.data.settings;
 	/** @type {Array<import('$lib/types').Group>} */
 	const groups = $page.data.groups;
+	/** @type {string} */
+	const runnerBackend = $page.data.runnerBackend;
 
 	$: userGroups = user.group_ids
 		.map((id) => groups.filter((g) => g.id === id)[0])
@@ -33,7 +37,7 @@
 </a>
 
 <div class="row mt-4">
-	<div class="col-md-8 col-lg-6">
+	<div class="col-md-10 col-lg-8">
 		<table class="table">
 			<tbody>
 				<tr>
@@ -61,27 +65,6 @@
 					<td><BooleanIcon value={user.is_verified} /></td>
 				</tr>
 				<tr>
-					<th>SLURM user</th>
-					<td>{user.slurm_user || '-'}</td>
-				</tr>
-				<tr>
-					<th>SLURM accounts</th>
-					<td>
-						{#if user.slurm_accounts.length > 0}
-							{#each user.slurm_accounts as account}
-								<span class="badge text-bg-light fw-normal fs-6">{account}</span>
-								&nbsp;
-							{/each}
-						{:else}
-							-
-						{/if}
-					</td>
-				</tr>
-				<tr>
-					<th>Cache dir</th>
-					<td>{user.cache_dir || '-'}</td>
-				</tr>
-				<tr>
 					<th>Groups</th>
 					<td>
 						{#each userGroups as group}
@@ -91,5 +74,60 @@
 				</tr>
 			</tbody>
 		</table>
+
+		{#if runnerBackend === 'slurm' || runnerBackend === 'slurm_ssh'}
+			<h4 class="fw-light ms-2 mt-4">Settings</h4>
+			<table class="table">
+				<tbody>
+					{#if runnerBackend === 'slurm'}
+						<tr>
+							<th>SLURM user</th>
+							<td>{settings.slurm_user || '-'}</td>
+						</tr>
+						<tr>
+							<th>Cache dir</th>
+							<td>{settings.cache_dir || '-'}</td>
+						</tr>
+					{/if}
+					{#if runnerBackend === 'slurm_ssh'}
+						<tr>
+							<th>SSH host</th>
+							<td>{settings.ssh_host || '-'}</td>
+						</tr>
+						<tr>
+							<th>SSH username</th>
+							<td>{settings.ssh_username || '-'}</td>
+						</tr>
+						<tr>
+							<th>SSH Private Key Path</th>
+							<td>{settings.ssh_private_key_path || '-'}</td>
+						</tr>
+						<tr>
+							<th>SSH Tasks Dir</th>
+							<td>{settings.ssh_tasks_dir || '-'}</td>
+						</tr>
+						<tr>
+							<th>SSH Jobs Dir</th>
+							<td>{settings.ssh_jobs_dir || '-'}</td>
+						</tr>
+					{/if}
+					{#if runnerBackend === 'slurm' || runnerBackend === 'slurm_ssh'}
+						<tr>
+							<th>SLURM accounts</th>
+							<td>
+								{#if settings.slurm_accounts.length > 0}
+									{#each settings.slurm_accounts as account}
+										<span class="badge text-bg-light fw-normal fs-6">{account}</span>
+										&nbsp;
+									{/each}
+								{:else}
+									-
+								{/if}
+							</td>
+						</tr>
+					{/if}
+				</tbody>
+			</table>
+		{/if}
 	</div>
 </div>